Move ObjCBuiltInFunctions to its own file and its singleton to data.dart
