+++ lib/src/code_generator.dart
@@ -16,6 +16,7 @@ export 'code_generator/handle.dart';
 export 'code_generator/imports.dart';
 export 'code_generator/library.dart';
 export 'code_generator/native_type.dart';
+export 'code_generator/objc_built_in_functions.dart';
 export 'code_generator/objc_interface.dart';
 export 'code_generator/pointer.dart';
 export 'code_generator/struct.dart';
