+++ lib/src/header_parser/data.dart
@@ -4,7 +4,8 @@
 
 import 'dart:ffi';
 
-import 'package:ffigen/src/code_generator.dart' show Constant;
+import 'package:ffigen/src/code_generator.dart'
+    show Constant, ObjCBuiltInFunctions;
 import 'package:ffigen/src/config_provider.dart' show Config;
 import 'clang_bindings/clang_bindings.dart' show Clang;
 
@@ -36,6 +37,10 @@ Map<String, Macro> _savedMacros = {};
 List<Constant> get unnamedEnumConstants => _unnamedEnumConstants;
 List<Constant> _unnamedEnumConstants = [];
 
+/// Built in functions used by the Objective C bindings.
+ObjCBuiltInFunctions get objCBuiltInFunctions => _objCBuiltInFunctions;
+late ObjCBuiltInFunctions _objCBuiltInFunctions;
+
 void initializeGlobals({required Config config}) {
   _config = config;
   _clang = Clang(DynamicLibrary.open(config.libclangDylib));
@@ -43,4 +48,5 @@ void initializeGlobals({required Config config}) {
   _savedMacros = {};
   _unnamedEnumConstants = [];
   _bindingsIndex = BindingsIndex();
+  _objCBuiltInFunctions = ObjCBuiltInFunctions();
 }
